+++ tests/test_rise.py
@@ -18,7 +18,7 @@ class RiseOnImages(TestCase):
 
     def test_rise_filename(self):
         model_filename = 'tests/test_data/mnist_model.onnx'
-        input_data = generate_data(batch_size=1)
+        input_data = generate_data(batch_size=1).astype(np.float32)
 
         heatmaps = dianna.explain_image(model_filename, input_data, method="RISE", n_masks=200)
